chore: moved OutputOptions to separate file
